Actualizacion y correccion de aplicacion Ajuste a nombres en un solo idioma de variables Ajuste a Exportacion de datos Ajuste a organizacion de componentes
