+++ app/Exports/MultiplesHojasExporte.php
@@ -5,16 +5,16 @@ namespace App\Exports;
 use Maatwebsite\Excel\Concerns\Exportable;
 use Maatwebsite\Excel\Concerns\WithMultipleSheets;
 
-class MultiSheetExport implements WithMultipleSheets
+class MultiplesHojasExporte implements WithMultipleSheets
 {
     use Exportable;
 
     public function sheets(): array
     {
         return [
-            new ClienteSheetExport(),
-            new GanadorSheetExport(),
-            new PremioSheetExport(),
+            new ClienteExporte,
+            new GanadorExporte(),
+            new PremioExporte()
         ];
     }
 }
