+++ app/Http/Controllers/ExportarDatosController.php
@@ -2,9 +2,8 @@
 
 namespace App\Http\Controllers;
 
-use App\Exports\MultiSheetExport;
+use App\Exports\MultiplesHojasExporte;
 use App\Http\Controllers\Controller;
-use Illuminate\Http\Request;
 use Maatwebsite\Excel\Facades\Excel;
 
 class ExportarDatosController extends Controller
@@ -13,6 +12,6 @@ class ExportarDatosController extends Controller
      * Función para exportar todas las tablas a un archivo xlsx.
      */
     public function exportarBase(){
-        return Excel::download(new MultiSheetExport, 'datos-sistema.xlsx');
+        return Excel::download(new MultiplesHojasExporte, 'datos-sistema.xlsx');
     }
 }
