+++ app/Http/Controllers/GanadorController.php
@@ -11,6 +11,8 @@ class GanadorController extends Controller
 {
     /**
      * Seleccionar ganador al azar si hay al menos 5 usuarios.
+     * 
+     * @return \Illuminate\Http\RedirectResponse
      */
     public function seleccionarGanador(){
         $usuarios = Cliente::all();
