+++ app/Http/Requests/UsuarioRequest.php
@@ -1,0 +1,33 @@
+<?php
+
+namespace App\Http\Requests;
+
+use Illuminate\Foundation\Http\FormRequest;
+use Illuminate\Validation\Rules;
+
+class UsuarioRequest extends FormRequest
+{
+    /**
+     * Determine if the user is authorized to make this request.
+     */
+    public function authorize(): bool
+    {
+        return true;
+    }
+
+    /**
+     * Get the validation rules that apply to the request.
+     *
+     * @return array<string, \Illuminate\Contracts\Validation\ValidationRule|array<mixed>|string>
+     */
+    public function rules(): array
+    {
+        return [
+            'identificacion' => ['required', 'string', 'max:15', 'unique:usuarios'],
+            'nombres' => ['required', 'string', 'max:45'],
+            'apellidos' => ['required', 'string', 'max:45'],
+            'correo' => ['required', 'email', 'max:100'],
+            'password' => ['required', 'confirmed', Rules\Password::defaults()],
+        ];
+    }
+}
