+++ resources/js/app.js
@@ -5,3 +5,5 @@ import Alpine from 'alpinejs';
 window.Alpine = Alpine;
 
 Alpine.start();
+
+import './municipios.js';
