+++ resources/views/layouts/guest.blade.php
@@ -27,7 +27,5 @@
                 {{ $slot }}
             </main>
         </div>
-
-        @yield('scripts')
     </body>
 </html>
